:art: refactor: 폴더구조 재구성
